添加多进程服务器

+++ src/Server/Server_Block.cpp
@@ -136,7 +136,7 @@ int server_block(int argc, char **argv)
 
     
 
-    int port = 0;
+    ssize_t port = 0;
     int backlog = 0; // 最大监听队列长度
     const char *ip = Prameter(argc,argv,port,backlog);
